refactor: organização de classes em modulos cenario, gridfrontend e personagem
