+++ src/main/java/ufpa/icen/pvz/view/GridFrontend/GridFrontend.java
@@ -1,5 +1,5 @@
 // arquivo: ufpa.icen.pvz.grid.Grid.java
-package ufpa.icen.pvz.view;
+package ufpa.icen.pvz.view.GridFrontend;
 
 /**
  * Classe auxiliar que representa o tabuleiro (grid) e faz conversões.
