+++ src/main/java/ufpa/icen/pvz/view/GridFrontend/GridPreset.java
@@ -1,5 +1,5 @@
 // arquivo: ufpa.icen.pvz.grid.GridPreset.java
-package ufpa.icen.pvz.view;
+package ufpa.icen.pvz.view.GridFrontend;
 
 /**
  * Presets para o tamanho da célula e valores padrão de grid.
