+++ src/main/java/ufpa/icen/pvz/view/cenario/MenuCenario.java
@@ -1,4 +1,4 @@
-package ufpa.icen.pvz.view;
+package ufpa.icen.pvz.view.cenario;
 import javax.swing.JButton;
 import javax.swing.ImageIcon;
 import javax.swing.KeyStroke;
